AoC 2024 Day 02
Part A tries: 1
Part B tries: 2

Tried to be too clever in the second half, reverted to brute forcing
when brute forcing as a double check ran in reasonable time.

I was discarding reports where [ an initial or final difference was too
large and wasn't fixed by dropping the first/last level ]. I did this
because I spotted a report that was not so fixed and then was considered
for the [ single wrong-way level ] checking even though it wouldn't have
passed that either. However, sometimes not so fixed reports are fixed
by dropping a wrong-way level in the second/second-last position.