Added new map dump
Strings are immutable so if you have an output table tracking an input
table, the output table has to be a list of a list of characters
instead of a list of strings. This requires a join() in the dumping to
make it look the same as the list of strings case.